checkout&payment was fixed
