--- frontend/src/components/CartModal/index.ts
@@ -1,1 +1,0 @@
-export * from './CartModal';
