+++ frontend/src/components/Filter/index.ts
@@ -1,0 +1,1 @@
+export * from './Filter';
