--- frontend/src/components/FilterModal/index.ts
@@ -1,1 +1,0 @@
-export * from './FilterModal';
