+++ frontend/src/components/LoginModal/LoginModal.scss
@@ -1,19 +1,11 @@
 @import "../../styles/main.scss";
 
 .LoginModal {
-  background-color: rgba(0, 0, 0, 0.5);
-  width: 100vw;
-  height: 100vh;
-  z-index: 1;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  position: fixed;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  &__container {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+  
     width: 832px;
     height: 542px;
     background-color: $main-background;
@@ -26,7 +18,6 @@
       width: 648px;
       height: 460px;
     }
-  }
 
   &__header {
     width: 100%;
