--- frontend/src/components/SearchModal/index.ts
@@ -1,1 +1,0 @@
-export * from './SearchModal';
