+++ frontend/src/contants/colors.ts
@@ -24,7 +24,7 @@ export const PRODUCT_HEX: { [color:string]: string } = {
 
 export const PAYMENT_STATUS_HEX: { [color:string]: string } = {
   "awaiting payment": '#DDD0C8',  
-  "payment recieved": '#B3BDA2',
+  "payment received": '#B3BDA2',
   "shipped": '#B3BDA2',
   "completed": '#B3BDA2',
   "refunded": '#DDD0C8',
