+++ frontend/src/pages/OrderCompleted/OrderCompleted.scss
@@ -4,5 +4,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  text-align: center;
   gap: 48px;
 }
