+++ frontend/src/pages/PhilosophyPage/PhilosophyPage.scss
@@ -5,7 +5,7 @@
   display: flex;
   justify-content: flex-end;
   gap: 176px;
-  padding-top: 32px;
+  padding: 32px 0 80px;
 
   @include onBigDesktop {
     gap: 100px;
